Сборщик проектов Мавен Создание и прочтение json файла Работы с REST API запросом
