+++ src/main/java/Garage/Main.java
@@ -1,7 +1,5 @@
-package JsonGarageTest;
+package Garage;
 
-import JsonGarageTest.Car;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.File;
