+++ src/main/java/WeatherTest/BaseParams.java
@@ -1,4 +1,4 @@
-package AccuweatherTest;
+package WeatherTest;
 
 public class BaseParams {
